+++ src/components/root/DiffQueueModals.tsx
@@ -3,7 +3,7 @@ import { useAppDispatch } from '../../state/store';
 import { selectNextForDiff } from '../../state/tabs/selectors';
 import { DialogTitle, Divider, Modal, ModalClose, ModalDialog } from '@mui/joy';
 import { popDiffQueue } from '../../state/tabs/slice';
-import { ResponseDiffOverlay } from './overlays/ResponseDiffOverlay';
+import { ResponseDiffOverlay } from './overlays/ResponseDiffOverlay/ResponseDiffOverlay';
 
 export function DiffQueueModals() {
 	const nextForDiff = useSelector(selectNextForDiff);
